requirements.txt
annotated-types==0.5.0
anyio==3.7.1
bcrypt==4.0.1
click==8.1.7
colorama==0.4.6
exceptiongroup==1.1.3
fastapi==0.103.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
jose==1.0.0
numpy==1.25.2
pandas==2.1.0
passlib==1.7.4
pydantic==1.10.12
pydantic-sqlalchemy==0.0.9
pydantic_core==2.6.3
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.49
starlette==0.27.0
typing_extensions==4.7.1
tzdata==2023.3
uvicorn==0.23.2
